# caresoft_core/caresoft_core/Connected Services/CoreWebApi/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "caresoft_core",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Factura/add": {
      "post": {
        "tags": [
          "Factura"
        ],
        "operationId": "add6",
        "parameters": [
          {
            "name": "FacturaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdCuenta",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ConsultaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdIngreso",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DocumentoCajero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MontoSubtotal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MontoTotal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/update": {
      "put": {
        "tags": [
          "Factura"
        ],
        "operationId": "update6",
        "parameters": [
          {
            "name": "FacturaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdCuenta",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ConsultaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdIngreso",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdSucursal",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "DocumentoCajero",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MontoSubtotal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MontoTotal",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Fecha",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/delete/{facturaCodigo}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "operationId": "delete6",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Factura/addFacturaServicio": {
      "post": {
        "tags": [
          "Factura"
        ],
        "operationId": "addFacturaServicio",
        "parameters": [
          {
            "name": "FacturaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ServicioCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdAutorizacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Resultados",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Costo",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/deleteFacturaServicio/{facturaCodigo}/{servicioCodigo}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "operationId": "deleteFacturaServicio",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "servicioCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/getFacturaServicios/{facturaCodigo}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "operationId": "getFacturaServicios",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServicioDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/addFacturaProducto": {
      "post": {
        "tags": [
          "Factura"
        ],
        "operationId": "addFacturaProducto",
        "parameters": [
          {
            "name": "FacturaCodigo",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IdProducto",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IdAutorizacion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Resultados",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Costo",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/deleteFacturaProducto/{facturaCodigo}/{idProducto}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "operationId": "deleteFacturaProducto",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idProducto",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/getFacturaProductos/{facturaCodigo}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "operationId": "getFacturaProductos",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaProductoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaProductoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacturaProductoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/addFacturaMetodoPago/{facturaCodigo}/{idMetodoPago}": {
      "post": {
        "tags": [
          "Factura"
        ],
        "operationId": "addFacturaMetodoPago",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idMetodoPago",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/deleteFacturaMetodoPago/{facturaCodigo}/{idMetodoPago}": {
      "delete": {
        "tags": [
          "Factura"
        ],
        "operationId": "deleteFacturaMetodoPago",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idMetodoPago",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Factura/getMetodoPagos/{facturaCodigo}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "operationId": "getMetodoPagos",
        "parameters": [
          {
            "name": "facturaCodigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodoPagoDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodoPagoDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetodoPagoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FacturaProductoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "facturaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "idAutorizacion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resultados": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "MetodoPagoDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idMetodoPago": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ServicioDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "servicioCodigo": {
            "type": "string",
            "nullable": true
          },
          "idTipoServicio": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}